ref(scheduler): move k8s scheduler to the top level of the module
